Refacor the database to implement the select term
Seperated the the terms in the database registration portion. Now the course detail and map link should work correctly
Also refactor the UI of the course detaled view to make the information display properly
Added term selection in the course list (Second Fragment) to seperate the courses based on the terms selected.

Co-Authored-By: Jiaqi Guo <422532105@users.noreply.github.com>