modified the location of the toast and background image
Show the toast that the Add or Drop class operation is successful only if the class is actually being added or dopped in the firebase database or display nothing otherwise.
The background image of the main activity(LoginInterface) is re-shooted and resized to correspond to the aspect ratio of the phones.