Refactored the UI
Added actionbars to major activties based on a toolbar implementation.
Modified the details of the application, including optimizing the positions and size of the ui components on some of the acitivies.
Added a comfirmation dialog at the Home "Page" so that the user won't be able to press back to go to the Login after they are signed in.
Added or modified margins and paddings of some UI elements(especially textviews) so that the Interface looks better.
-Anran