PreDemo Improvements
Some small detail fixes including: Added the dialog for every button that can be continuously pressed so that the app won't be overloaded with too much requests from the user. Also that helps to prevent the UI from not reponding if the clicks are being made too frequently.
Added limitations and null checks at the add drop class so that the nuccessary queries to the database can be cut off locally to reduce traffic and improve performance & efficiency.
@jbv777 Some more tweaking of the timetable